1107 Move CIDR/IPFilterRules to protocol package; fix CIDR mask handling
Move CIDR and IPFilterRules from the okhttp package to
org.apache.stormcrawler.protocol so they can be reused by other
protocols (review feedback).

Fix CIDR byte mask wrapping for /32, /128 and IPv6 prefixes >= 32
(Java shifts mod 32) and reject out-of-range masks. Add a NOTE that
IP filtering is bypassed for proxied fetches.
